OCR_sim/requirements.txt

Flask
Pillow
pytesseract
python-Levenshtein